［需求／BUG］: 优化菜品状态更新代码 ［模块］: ［说明］: ［作者］: 桥泸007

+++ sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -100,7 +100,7 @@ public class DishController {
      */
     @PostMapping("/status/{status}")
     @ApiOperation("菜品起售、停售")
-    public Result updateStatus(Long id, @PathVariable Long status) {
+    public Result updateStatus(Long id, @PathVariable Integer status) {
         log.info("菜品起售、停售：id {}，status {}", id, status);
         dishService.updateStatus(id, status);
         return Result.success();
